+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/source/ads/AdsMediaSource.java
@@ -59,13 +59,12 @@ import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
+import org.checkerframework.checker.nullness.qual.RequiresNonNull;
 
 /**
  * A {@link MediaSource} that inserts ads linearly into a provided content media source.
  *
- * <p>The wrapped content media source can contain multiple {@linkplain Timeline.Period periods}.
- * Note that multi-period ad insertion is not supported by all {@link AdsLoader} implementations
- * (for example, the IMA extension's {@code ImaAdsLoader} only supports single-period Timelines).
+ * <p>The wrapped content media source must contain a single {@link Timeline.Period}.
  */
 @UnstableApi
 public final class AdsMediaSource extends CompositeMediaSource<MediaPeriodId> {
@@ -154,12 +153,11 @@ public final class AdsMediaSource extends CompositeMediaSource<MediaPeriodId> {
   private final boolean useLazyContentSourcePreparation;
   private final boolean useAdMediaSourceClipping;
   private final List<AdMediaSourceHolder> activeMediaSourceHolders;
-  private final Map<ClippingMediaPeriod, MediaPeriodId> activeContentClippingMediaPeriods;
+  private final Map<ClippingMediaPeriod, Integer> activeContentClippingMediaPeriods;
 
   // Accessed on the player thread.
   @Nullable private ComponentListener componentListener;
   @Nullable private Timeline contentTimeline;
-  @Nullable private Timeline activeTimeline;
   @Nullable private AdPlaybackState adPlaybackState;
   private @NullableType AdMediaSourceHolder[][] adMediaSourceHolders;
   @Nullable private Handler playerHandler;
@@ -332,16 +330,14 @@ public final class AdsMediaSource extends CompositeMediaSource<MediaPeriodId> {
       if (id.nextAdGroupIndex == C.INDEX_UNSET) {
         return maskingMediaPeriod;
       }
-      long endPositionUs =
-          getContentClippingEndPositionUs(
-              checkNotNull(activeTimeline), id.periodUid, id.nextAdGroupIndex);
+      long nextAdGroupTimeUs = adPlaybackState.getAdGroup(id.nextAdGroupIndex).timeUs;
       ClippingMediaPeriod clippingMediaPeriod =
           new ClippingMediaPeriod(
               maskingMediaPeriod,
               /* enableInitialDiscontinuity= */ true,
               /* startUs= */ 0,
-              /* endUs= */ endPositionUs);
-      activeContentClippingMediaPeriods.put(clippingMediaPeriod, id);
+              /* endUs= */ nextAdGroupTimeUs);
+      activeContentClippingMediaPeriods.put(clippingMediaPeriod, id.nextAdGroupIndex);
       return clippingMediaPeriod;
     }
   }
@@ -379,7 +375,6 @@ public final class AdsMediaSource extends CompositeMediaSource<MediaPeriodId> {
     this.playerHandler = null;
     componentListener.stop();
     contentTimeline = null;
-    activeTimeline = null;
     adPlaybackState = null;
     adMediaSourceHolders = new AdMediaSourceHolder[0][];
     mainHandler.post(() -> adsLoader.stop(/* adsMediaSource= */ this, componentListener));
@@ -395,6 +390,7 @@ public final class AdsMediaSource extends CompositeMediaSource<MediaPeriodId> {
           .handleSourceInfoRefresh(newTimeline);
       maybeUpdateSourceInfo();
     } else {
+      checkArgument(newTimeline.getPeriodCount() == 1);
       contentTimeline = newTimeline;
       mainHandler.post(
           () -> {
@@ -448,6 +444,14 @@ public final class AdsMediaSource extends CompositeMediaSource<MediaPeriodId> {
             adMediaSourceHolder.setEndPositionUs(adDurationUs);
           }
         }
+      }
+      for (Map.Entry<ClippingMediaPeriod, Integer> activeClippingPeriod :
+          activeContentClippingMediaPeriods.entrySet()) {
+        int nextAdGroupIndex = activeClippingPeriod.getValue();
+        long nextAdGroupTimeUs = adPlaybackState.getAdGroup(nextAdGroupIndex).timeUs;
+        activeClippingPeriod
+            .getKey()
+            .updateClipping(/* startUs= */ 0, /* endUs= */ nextAdGroupTimeUs);
       }
     }
     this.adPlaybackState = adPlaybackState;
@@ -532,30 +536,19 @@ public final class AdsMediaSource extends CompositeMediaSource<MediaPeriodId> {
 
   private void maybeUpdateSourceInfo() {
     @Nullable Timeline contentTimeline = this.contentTimeline;
-    @Nullable AdPlaybackState adPlaybackState = this.adPlaybackState;
     if (adPlaybackState != null && contentTimeline != null) {
       if (adPlaybackState.adGroupCount == 0) {
-        activeTimeline = contentTimeline;
         refreshSourceInfo(contentTimeline);
       } else {
-        long[][] durations = getAdDurationsUs(adPlaybackState);
-        adPlaybackState = adPlaybackState.withAdDurationsUs(durations);
-        this.adPlaybackState = adPlaybackState;
-        Timeline activeTimeline = new AdTimeline(contentTimeline, adPlaybackState);
-        this.activeTimeline = activeTimeline;
-        for (Map.Entry<ClippingMediaPeriod, MediaPeriodId> activeClippingPeriod :
-            activeContentClippingMediaPeriods.entrySet()) {
-          MediaPeriodId id = activeClippingPeriod.getValue();
-          long endPositionUs =
-              getContentClippingEndPositionUs(activeTimeline, id.periodUid, id.nextAdGroupIndex);
-          activeClippingPeriod.getKey().updateClipping(/* startUs= */ 0, endPositionUs);
-        }
-        refreshSourceInfo(activeTimeline);
+        adPlaybackState = adPlaybackState.withAdDurationsUs(getAdDurationsUs());
+        refreshSourceInfo(new SinglePeriodAdTimeline(contentTimeline, adPlaybackState));
       }
     }
   }
 
-  private long[][] getAdDurationsUs(AdPlaybackState adPlaybackState) {
+  @RequiresNonNull("adPlaybackState")
+  private long[][] getAdDurationsUs() {
+    AdPlaybackState adPlaybackState = checkNotNull(this.adPlaybackState);
     boolean hasPostRollPlaceholder = adPlaybackState.endsWithLivePostrollPlaceHolder();
     int adGroupCount = adMediaSourceHolders.length + (hasPostRollPlaceholder ? 1 : 0);
     long[][] adDurationsUs = new long[adGroupCount][];
@@ -805,15 +798,5 @@ public final class AdsMediaSource extends CompositeMediaSource<MediaPeriodId> {
               ? ((ClippingMediaPeriod) mediaPeriod).mediaPeriod
               : mediaPeriod);
     }
-  }
-
-  private long getContentClippingEndPositionUs(
-      Timeline activeTimeline, Object periodUid, int nextAdGroupIndex) {
-    int periodIndex = activeTimeline.getIndexOfPeriod(periodUid);
-    if (periodIndex == C.INDEX_UNSET) {
-      return C.TIME_END_OF_SOURCE;
-    }
-    activeTimeline.getPeriod(periodIndex, period);
-    return period.adPlaybackState.getAdGroup(nextAdGroupIndex).timeUs;
   }
 }