+++ libraries/exoplayer/src/test/java/androidx/media3/exoplayer/ExoPlayerAdPlaybackEnforcementTest.java
@@ -237,7 +237,9 @@ public final class ExoPlayerAdPlaybackEnforcementTest {
     advance(player).untilTimelineChanges();
     AdPlaybackState updatedAdPlaybackState =
         emptyAdPlaybackState
-            .withNewAdGroup(/* adGroupIndex= */ 0, /* adGroupTimeUs= */ 0)
+            .withNewAdGroup(
+                /* adGroupIndex= */ 0,
+                /* adGroupTimeUs= */ DEFAULT_WINDOW_OFFSET_IN_FIRST_PERIOD_US)
             .withAdCount(/* adGroupIndex= */ 0, /* adCount= */ 1)
             .withAvailableAdMediaItem(
                 /* adGroupIndex= */ 0,
@@ -379,7 +381,7 @@ public final class ExoPlayerAdPlaybackEnforcementTest {
     TestExoPlayerBuilder playerBuilder =
         new TestExoPlayerBuilder(ApplicationProvider.getApplicationContext());
     ExoPlayer player = playerBuilder.setEnforceAdPlaybackOnTimelineRefresh(false).build();
-    long firstSampleTimeUs = 0;
+    long firstSampleTimeUs = DEFAULT_WINDOW_OFFSET_IN_FIRST_PERIOD_US;
     AdPlaybackState initialAdPlaybackState =
         new AdPlaybackState("adsId", firstSampleTimeUs)
             .withAvailableAdMediaItem(
@@ -463,7 +465,7 @@ public final class ExoPlayerAdPlaybackEnforcementTest {
     TestExoPlayerBuilder playerBuilder =
         new TestExoPlayerBuilder(ApplicationProvider.getApplicationContext());
     ExoPlayer player = playerBuilder.setEnforceAdPlaybackOnTimelineRefresh(false).build();
-    long firstSampleTimeUs = 0;
+    long firstSampleTimeUs = DEFAULT_WINDOW_OFFSET_IN_FIRST_PERIOD_US;
     AdPlaybackState initialAdPlaybackState =
         new AdPlaybackState("adsId", firstSampleTimeUs + 1_000_000L)
             .withAvailableAdMediaItem(
@@ -634,7 +636,7 @@ public final class ExoPlayerAdPlaybackEnforcementTest {
     TestExoPlayerBuilder playerBuilder =
         new TestExoPlayerBuilder(ApplicationProvider.getApplicationContext());
     ExoPlayer player = playerBuilder.setEnforceAdPlaybackOnTimelineRefresh(false).build();
-    long firstSampleTimeUs = 0;
+    long firstSampleTimeUs = DEFAULT_WINDOW_OFFSET_IN_FIRST_PERIOD_US;
     AdPlaybackState initialAdPlaybackState =
         new AdPlaybackState("adsId", firstSampleTimeUs)
             .withAvailableAdMediaItem(