+++ libraries/exoplayer_ima/src/main/java/androidx/media3/exoplayer/ima/AdTagLoader.java
@@ -21,7 +21,6 @@ import static androidx.media3.exoplayer.ima.ImaUtil.BITRATE_UNSET;
 import static androidx.media3.exoplayer.ima.ImaUtil.TIMEOUT_UNSET;
 import static androidx.media3.exoplayer.ima.ImaUtil.getAdGroupTimesUsForCuePoints;
 import static androidx.media3.exoplayer.ima.ImaUtil.getImaLooper;
-import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.common.base.Preconditions.checkState;
 import static java.lang.Math.max;
@@ -491,14 +490,7 @@ import java.util.Objects;
     }
     Player player = this.player;
     this.timeline = timeline;
-    int windowIndex = player.getCurrentMediaItemIndex();
-    Timeline.Window window = new Timeline.Window();
-    timeline.getWindow(windowIndex, window);
-    checkArgument(
-        window.firstPeriodIndex == window.lastPeriodIndex,
-        "Only single-period Timelines are supported.");
-    long contentDurationUs = timeline.getPeriod(window.firstPeriodIndex, period).durationUs;
-
+    long contentDurationUs = timeline.getPeriod(player.getCurrentPeriodIndex(), period).durationUs;
     contentDurationMs = Util.usToMs(contentDurationUs);
     if (contentDurationUs != adPlaybackState.contentDurationUs) {
       adPlaybackState = adPlaybackState.withContentDurationUs(contentDurationUs);